split-out patches into individual files
