+++ patches/store_magic_on_flash.py
@@ -1,0 +1,6 @@
+def store_magic_on_flash__patch(maincpu_patches):
+    maincpu_patches[0xEF197C] = [
+        0x42, 0x00, 0x00, 0x3B, 0x00,   # ld XDE, 0x003B0000
+        0xF5, 0xE9, 0x02, 0x34, 0x12,   # ld (XDE+), 0x1234
+        0x00,  # nop
+    ]
